workaround gnutls internal buffering
https://bugzilla.xfce.org/show_bug.cgi?id=9284

xfce_mailwatch_net_conn_recv_internal always checks if the
fd is readable, but if gnutls is buffering further data from
the last packet internally, then the fd will not be readable
despite there being more data to return.

By making the buffer always big enough to receive a standard
sized internet packet, we are reducing the risk of this case
happening.  It's not a perfect solution, but it sure is
trivial...

(a lot more trivial than this longwinded commit message!)